+++ app/Adms/Views/pagetypes/pageTypes.php
@@ -1,12 +1,19 @@
 <link rel="stylesheet" href="<?= \Core\Config::url() . '/assets/css/pages/list.css' ?>">
 
+<?php
+/** @var array<string, bool> $buttonpermissions */
+$buttonPermissions = $buttonpermissions ?? [];
+?>
+
 <div class="list-page">
     <div class="list-page__header">
         <h2 class="list-page__title">Tipos de Página</h2>
         <div class="list-page__actions">
-            <a href="<?= \Core\Config::url() ?>/add-page-type/index" class="btn btn-primary">
-                <i class="fa-solid fa-plus"></i> Adicionar tipo
-            </a>
+            <?php if (!empty($buttonPermissions['add'])) : ?>
+                <a href="<?= \Core\Config::url() ?>/add-page-type/index" class="btn btn-primary">
+                    <i class="fa-solid fa-plus"></i> Adicionar tipo
+                </a>
+            <?php endif; ?>
         </div>
     </div>
 
@@ -33,16 +40,22 @@
                             <td><?= htmlspecialchars($type['order_page_type']) ?></td>
                             <td>
                                 <div class="list-table__actions">
-                                    <a href="<?= \Core\Config::url() ?>/view-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline">
-                                        <i class="fa-solid fa-eye"></i> Visualizar
-                                    </a>
-                                    <a href="<?= \Core\Config::url() ?>/update-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline">
-                                        <i class="fa-solid fa-pen"></i> Editar
-                                    </a>
-                                    <a href="<?= \Core\Config::url() ?>/delete-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline"
-                                        onclick="return confirm('Tem certeza que deseja excluir este tipo de página?');">
-                                        <i class="fa-solid fa-trash"></i> Excluir
-                                    </a>
+                                    <?php if (!empty($buttonPermissions['view'])) : ?>
+                                        <a href="<?= \Core\Config::url() ?>/view-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline">
+                                            <i class="fa-solid fa-eye"></i> Visualizar
+                                        </a>
+                                    <?php endif; ?>
+                                    <?php if (!empty($buttonPermissions['update'])) : ?>
+                                        <a href="<?= \Core\Config::url() ?>/update-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline">
+                                            <i class="fa-solid fa-pen"></i> Editar
+                                        </a>
+                                    <?php endif; ?>
+                                    <?php if (!empty($buttonPermissions['delete'])) : ?>
+                                        <a href="<?= \Core\Config::url() ?>/delete-page-type/index/<?= $type['id'] ?>" class="btn btn-sm btn-outline"
+                                            onclick="return confirm('Tem certeza que deseja excluir este tipo de página?');">
+                                            <i class="fa-solid fa-trash"></i> Excluir
+                                        </a>
+                                    <?php endif; ?>
                                 </div>
                             </td>
                         </tr>
